fix: not logging error on 400 + reorganizing + tests
