+++ iaso/urls.py
@@ -110,7 +110,7 @@ from .api.permissions.permissions import PermissionsViewSet
 from .api.profiles.views import ProfilesViewSet
 from .api.projects import ProjectsViewSet
 from .api.reports import ReportsViewSet
-from .api.setup_account import SetupAccountViewSet
+from .api.setup_account.views import SetupAccountViewSet
 from .api.source_versions import SourceVersionViewSet
 from .api.stocks.views import (
     StockItemRuleViewSet,
